+++ packages/@expo/cli/e2e/__tests__/customize-test.ts
@@ -64,14 +64,10 @@ it(
   'runs `npx expo customize`',
   async () => {
     const projectRoot = await setupTestProjectAsync('basic-customize', 'with-blank');
-    // `npx expo customize index.html serve.json babel.config.js`
-    await execa(
-      'node',
-      [bin, 'customize', 'public/index.html', 'public/serve.json', 'babel.config.js'],
-      {
-        cwd: projectRoot,
-      }
-    );
+    // `npx expo customize index.html babel.config.js`
+    await execa('node', [bin, 'customize', 'public/index.html', 'babel.config.js'], {
+      cwd: projectRoot,
+    });
 
     const files = klawSync(projectRoot)
       .map((entry) => {
@@ -89,7 +85,6 @@ it(
       'bun.lockb',
       'package.json',
       'public/index.html',
-      'public/serve.json',
     ]);
   },
   // Could take 45s depending on how fast npm installs